+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/querychannels/ChannelFilterRequest.kt
@@ -7,7 +7,6 @@ import io.getstream.chat.android.client.call.await
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Filters
 import io.getstream.chat.android.client.utils.Result
-import kotlinx.coroutines.flow.StateFlow
 
 /* Default filter to include FilterObject in a channel by its cid
 *
@@ -29,27 +28,4 @@ internal object ChannelFilterRequest {
                 memberLimit = 0,
             )
         ).await()
-}
-
-/**
- * Checks if the channel collection contains a channel, if yes then it returns skip handling result, otherwise it
- * adds the channel.
- */
-internal fun addIfChannelIsAbsent(channels: StateFlow<List<Channel>>, channel: Channel?): EventHandlingResult {
-    return if (channel == null || channels.value.any { it.cid == channel.cid }) {
-        EventHandlingResult.Skip
-    } else {
-        EventHandlingResult.Add(channel)
-    }
-}
-
-/**
- * Checks if the channel collection contains a channel, if yes then it removes it. Otherwise it simply skips the event
- */
-internal fun removeIfChannelIsPresent(channels: StateFlow<List<Channel>>, channel: Channel?): EventHandlingResult {
-    return if (channel != null && channels.value.any { it.cid == channel.cid }) {
-        EventHandlingResult.Remove(channel.cid)
-    } else {
-        EventHandlingResult.Skip
-    }
 }
